added some new packages to the exceptions, next step is to test some validators. Will test after pull request
