+++ src/main/java/com/splitface/tattoo/exception/GlobalExceptionHandler.java
@@ -1,8 +1,8 @@
 package com.splitface.tattoo.exception;
 
 
-import com.splitface.tattoo.exception.exceptions.*;
-import org.apache.coyote.Response;
+import com.splitface.tattoo.exception.artistCheck.artistCheckExceptions.*;
+import com.splitface.tattoo.exception.serviceLayer.serviceLayerExceptions.*;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
@@ -11,62 +11,10 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 @ControllerAdvice
 public class GlobalExceptionHandler {
 
-    @ExceptionHandler
-    public ResponseEntity<Object> handleEmptyArtistTableException(EmptyArtistTableException e){
-        return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_FOUND);
-    }
-  
-    @ExceptionHandler
-    public ResponseEntity<Object> handleEmptyTattooTableException(EmptyTattooTableException e){
-        return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_FOUND);
-    }
-    @ExceptionHandler
-    public ResponseEntity<Object> handlePasswordValidatorException(PasswordValidatorException e){
-        return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
-    }
-    @ExceptionHandler
-    public ResponseEntity<Object> handleNameValidatorException(NameValidatorException e) {
-        return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
-    }
-    @ExceptionHandler
-    public ResponseEntity<Object> handleMatchingStyleIdException(TattooMatchingStyleIdException e){
-        return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
-    }
 
-    @ExceptionHandler
-    public ResponseEntity<Object> handleArtistIdException(ArtistIdDoesNotExistException e){
-        return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_FOUND);
-    }
 
-    @ExceptionHandler
-    public ResponseEntity<Object> handleEmptyStyleTableException(EmptyStyleTableException e){
-        return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_FOUND);
-    }
 
-    @ExceptionHandler
-    public ResponseEntity<Object> handlePostcodeValidatorException(PostcodeValidatorException e){
-        return new ResponseEntity<>(e.getMessage(),HttpStatus.NOT_ACCEPTABLE);
-    }
 
-    @ExceptionHandler
-    public ResponseEntity<Object> handleEmailValidatorException(EmailValidatorException e){
-        return new ResponseEntity<>(e.getMessage(),HttpStatus.NOT_ACCEPTABLE);
-    }
-
-    @ExceptionHandler
-    public ResponseEntity<Object> handleTattooIdDoesNotExistException(TattooIdDoesNotExistException e){
-        return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_FOUND);
-    }
-
-    @ExceptionHandler
-    public ResponseEntity<Object> handleInvalidEmailException(InvalidEmailException e){
-        return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_ACCEPTABLE);
-    }
-
-    @ExceptionHandler
-    public ResponseEntity<Object> handleInvalidPostcodeException(InvalidPostcodeException e){
-        return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_ACCEPTABLE);
-    }
