+++ src/main/java/com/splitface/tattoo/exception/artistCheck/artistCheckExceptions/EmailValidatorException.java
@@ -1,4 +1,4 @@
-package com.splitface.tattoo.exception.exceptions;
+package com.splitface.tattoo.exception.artistCheck.artistCheckExceptions;
 
 public class EmailValidatorException extends RuntimeException{
     public EmailValidatorException(String message) {
