+++ src/main/java/com/splitface/tattoo/exception/artistCheck/artistCheckExceptions/InvalidEmailException.java
@@ -1,4 +1,4 @@
-package com.splitface.tattoo.exception.exceptions;
+package com.splitface.tattoo.exception.artistCheck.artistCheckExceptions;
 
 public class InvalidEmailException extends RuntimeException{
     public InvalidEmailException(String info){
