+++ src/main/java/com/splitface/tattoo/exception/artistCheck/artistCheckExceptions/PostcodeValidatorException.java
@@ -1,4 +1,4 @@
-package com.splitface.tattoo.exception.exceptions;
+package com.splitface.tattoo.exception.artistCheck.artistCheckExceptions;
 
 public class PostcodeValidatorException extends RuntimeException{
     public PostcodeValidatorException(String message) {
