--- src/main/java/com/splitface/tattoo/exception/exceptions/InvalidPostcodeException.java
@@ -1,8 +1,0 @@
-package com.splitface.tattoo.exception.exceptions;
-
-public class InvalidPostcodeException extends RuntimeException{
-
-    public InvalidPostcodeException(String info){
-        super(info);
-    }
-}
