+++ src/main/java/com/splitface/tattoo/exception/serviceLayer/serviceLayerExceptions/EmptyStyleTableException.java
@@ -1,4 +1,4 @@
-package com.splitface.tattoo.exception.exceptions;
+package com.splitface.tattoo.exception.serviceLayer.serviceLayerExceptions;
 
 public class EmptyStyleTableException extends RuntimeException{
     public EmptyStyleTableException(String message) {
