+++ src/main/java/com/splitface/tattoo/exception/serviceLayer/serviceLayerExceptions/TattooMatchingStyleIdException.java
@@ -1,4 +1,4 @@
-package com.splitface.tattoo.exception.exceptions;
+package com.splitface.tattoo.exception.serviceLayer.serviceLayerExceptions;
 
 public class TattooMatchingStyleIdException extends RuntimeException{
