+++ src/main/java/com/splitface/tattoo/models/Artist.java
@@ -3,10 +3,7 @@ package com.splitface.tattoo.models;
 import com.fasterxml.jackson.annotation.JsonBackReference;
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 import java.util.List;
 
@@ -14,6 +11,8 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
+@Getter
+@Setter
 @Table
 public class Artist {
 
@@ -30,52 +29,6 @@ public class Artist {
     @OneToMany(mappedBy = "artist", fetch = FetchType.LAZY)
     private List<Tattoo> tattoos;
 
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getLocation() {
-        return location;
-    }
-
-    public void setLocation(String location) {
-        this.location = location;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public List<Tattoo> getTattoos() {
-        return tattoos;
-    }
-
-    public void setTattoos(List<Tattoo> tattoos) {
-        this.tattoos = tattoos;
-    }
+    private String profileURL;
 }
