+++ src/main/java/com/splitface/tattoo/service/StyleService.java
@@ -7,4 +7,5 @@ import java.util.List;
 public interface StyleService {
     List<Style> getAllStylesFromDB();
     void addStylesForTattoo(Long tattooId, List<String> styleNames);
+
 }
