+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.profiles.active=prod
+spring.profiles.active=dev
 
 #spring.datasource.url=jdbc:postgresql://localhost:5432/tattoo_shop
 #spring.datasource.username=ion
