+++ src/test/java/com/splitface/tattoo/service/test/ArtistServiceImplTest.java
@@ -1,10 +1,10 @@
 package com.splitface.tattoo.service.test;
 
 
-import com.splitface.tattoo.exception.exceptions.ArtistIdDoesNotExistException;
+import com.splitface.tattoo.exception.serviceLayer.serviceLayerExceptions.ArtistIdDoesNotExistException;
 
-import com.splitface.tattoo.exception.exceptions.EmptyArtistTableException;
-import com.splitface.tattoo.exception.exceptions.PasswordValidatorException;
+import com.splitface.tattoo.exception.serviceLayer.serviceLayerExceptions.EmptyArtistTableException;
+import com.splitface.tattoo.exception.artistCheck.artistCheckExceptions.PasswordValidatorException;
 import com.splitface.tattoo.models.Artist;
 import com.splitface.tattoo.repository.ArtistRepository;
 import com.splitface.tattoo.service.ArtistServiceImpl;
